CalmrModel methods

S4 methods for CalmrModel

Usage

# S4 method for class 'CalmrModel'
run(object, experience, mapping, timings, ...)

# S4 method for class 'CalmrModel'
parameters(x)

# S4 method for class 'CalmrModel'
parameters(x) <- value

# S4 method for class 'CalmrModel'
raw_results(object)

# S4 method for class 'CalmrModel'
parsed_results(object)

# S4 method for class 'CalmrModel'
show(object)

# S4 method for class 'CalmrModel'
parse(object, outputs = object@outputs)

# S4 method for class 'CalmrModel'
plot(x, type = NULL, ...)

# S4 method for class 'CalmrModel'
graph(x, ...)

# S4 method for class 'ANCCR'
run(object, experience, mapping, timings, ..., debug = FALSE, debug_t = -1)

# S4 method for class 'HDI2020'
run(object, experience, mapping, ...)

# S4 method for class 'HD2022'
run(object, experience, mapping, ...)

# S4 method for class 'MAC1975'
run(object, experience, mapping, ...)

# S4 method for class 'PKH1982'
run(object, experience, mapping, ...)

# S4 method for class 'RAND'
run(object, experience, mapping, ...)

# S4 method for class 'RW1972'
run(object, experience, mapping, ...)

# S4 method for class 'SM2007'
run(
  object,
  experience,
  mapping,
  debug = FALSE,
  comparator_func = .witnauer_comparator_proc,
  ...
)

# S4 method for class 'TD'
run(object, experience, mapping, timings, ...)

Arguments

object

A CalmrModel object.

experience

A data.frame specifying trials as rows, as returned by make_experiment().

mapping

A named list specifying trial and stimulus mapping, as returned by make_experiment().

timings

A named list specifying timings for the model. Only used for timed models.

...

Additional named arguments.

x

A CalmrModel object.

value

A list of parameters to set.

outputs

A character vector specifying the outputs to parse. If not specified, all outputs of the model will be parsed.

type

A character vector specifying the types of plots to generate (should be model outputs).

debug

A logical to print debugging messages.

debug_t

A trial to debug at.

comparator_func

The function for the comparator process.

Value

run() returns the CalmrModel after running the phases in the design.

parameters() returns the parameters of the CalmrModel object.

parameters()<- sets the parameters of a CalmrModel object.

raw_results() returns the last raw results of the CalmrModel object.

parsed_results() returns the last parsed results of the CalmrModel object.

show() returns NULL (invisibly).

parse() returns CalmrModel with parsed results.

plot() returns a list of 'ggplot' plot objects.

graph() returns a 'ggplot' object.

Note

The run method changes some internal states of the model (if appropriate) and populates the .last_raw_results slot with the results of the run.