Perform representational similarity analysis on model comparison

Perform representational similarity analysis on model comparison

Usage

RSA(
  comparison,
  trials = "all",
  dist_method = "euclidean",
  corr_method = "spearman",
  ...
)

RSATest(comparison, n_samples = 1000, p = 0.95, ...)

Arguments

comparison

A CalmrComparison object, as returned by compare_models

trials

A character specifying what trials to use for the comparison. One of c("all", "last"). Default is "all". See details.

dist_method

A character specifying the distance method to be used in dist. Default is "euclidean".

corr_method

A character specifying the correlation method to be used in cor. Default is "spearman".

...

Additional parameters passed to dist and cor (in the case of RSA) or RSA (in the case of RSATest).

n_samples

An integer specifying the number of samples for the permutation test. Default is 1e3.

p

A float between 0 and 1 specifying the critical value of the null distribution to use for significance testing. Default is 0.95.

Value

A CalmrRSA or CalmrRSATest object.

Details

The trials argument specifyies which trials to use in the calculation of distance matrices; "all" means every trial will be used. "last" means only the last trial of each unique trial type will be used.