Train the SM2007 model

Train the SM2007 model

Usage

SM2007(
  alphas,
  lambdas,
  omegas,
  rhos,
  gammas,
  taus,
  orders,
  V = NULL,
  O = NULL,
  experience,
  mapping,
  debug = F,
  comparator_func = .witnauer_comparator_proc,
  ...
)

Arguments

alphas

A named vector with stimulus saliences.

lambdas

A named vector with learning asymptotes.

omegas

A named vector with weakening rates (extinction).

rhos

A named vector with weights for activation of present stimuli.

gammas

A named vector with weights for contribution of each stimulus in comparison processes.

taus

A named vector with stimulus-specific rates for operator switch.

orders

An integer specifying the order of the comparison process. Defaults to 1.

V

(optional) A named matrix of dimensions S,S (the associative strengths); where S is the number of stimuli.

O

(optional) A named matrix of dimensions S,S,S (the operator switches).

experience

A data.frame specifying trials as rows, as returned by make_model_args

mapping

A named list specifying trial and stimulus mapping, as returned by make_model_args

debug

A logical specifying whether to print information for the comparison process. Defaults to FALSE.

comparator_func

A function to be used in the comparison process. Either .witnauer_comparator_funcor .comparator_func.

Value

A list with

vs, Array of dimensions P,S,S; where P is the number of trials used to train the model and S is the number of stimuli involved in the experiment. Contains stimulus association strengths.

acts As above but with stimulus activations.

relacts As above but with (relative) stimulus activations after comparison process.

os Array of dimensions P,S,S,S, with the state of the operator switches. See notes related to O parameter.

other, Carryover arguments used for further processing by other functions

Note

Correct usage of the V and O parameters requires the matrices to adhere to a specific format. For V, entry i,j represents the association from stimulus i to stimulus j. For O, entry i,k,j represents the operator switch between stimulus i and comparator k with respect to stimulus j.