Make a tibble to fit a calmr model

Make a tibble to fit a calmr model

Usage

make_model_args(design, pars = NULL, model = NULL, opts = get_exp_opts())

Arguments

design

A design data.frame

pars

A data.frame containing parameters as returned by get_model_params

model

A string specifying the model name. One of supported_models()

opts

A list with options as returned by get_exp_opts

Value

A tibble with the arguments required to run the model. Each row represents a group in the experimental design.

See also

Examples

des <- data.frame(Group = "G1", P1 = "10A>(US)", R1 = TRUE)
ps <- get_model_params(des, model = "HD2022")
op <- get_exp_opts(iterations = 1)
make_model_args(design = des, pars = ps, model = "HD2022", opts = op)
#> # A tibble: 1 × 6
#>   model  iteration group experience    mapping          alphas   
#>   <chr>      <int> <chr> <list>        <list>           <list>   
#> 1 HD2022         1 G1    <df [10 × 5]> <named list [9]> <dbl [2]>