General plotting functions

plot_targetted_tbins() plots targetted time data on a trial.

plot_tbins() plots non-targetted time data on a trial.

plot_targetted_trials() plots targetted trial data.

plot_trials() plots non-targetted trial data.

plot_targetted_typed_trials() plots targetted trial data with a type.

plot_targetted_complex_trials() plots targetted data with a third variable.

Usage

plot_targetted_tbins(data, t = max(data$trial))

plot_tbins(data, t = max(data$trial))

plot_targetted_trials(data)

plot_trials(data)

plot_targetted_typed_trials(data)

plot_targetted_complex_trials(data, col)

Arguments

data

A data.frame-like with data to plot.

t

A numeric vector specifying the trial(s) to plot. Defaults to the last trial in data.

col

A string specifying the column of the third variable.

Value

plot_targetted_tbins() returns 'ggplot' object.

plot_tbins() returns 'ggplot' object.

plot_targetted_trials() returns 'ggplot' object.

plot_trials() returns 'ggplot' object.

plot_targetted_typed_trials() returns 'ggplot' object.

plot_targetted_complex_trials() returns 'ggplot' object.

Note

All data must be organised as returned by results() or parsed_results().