Fit model to data

Obtain MLE estimates for model, given data

Usage

fit_model(data, model_function, optimizer_options, file = NULL, ...)

Arguments

data

A numeric vector containing data to fit model against.

model_function

A function that runs the model and returns data.frame of r-values, organized as data.

optimizer_options

A list with options for the optimizer, as returned by get_optimizer_opts.

file

A path to save the model fit. If the arguments to the fit call are found to be identical to those in the file, the model just gets loaded.

...

Extra parameters passed to the optimizer call

Value

A CalmrFit object

Note

See the calmr_fits vignette for examples

See also